requirements.txt
fastapi==0.136.1
uvicorn[standard]==0.46.0
requests==2.33.1
requests-ntlm==1.3.0
requests-negotiate-sspi==0.5.2
pandas==3.0.2
openpyxl==3.1.5
python-dotenv==1.2.2
pydantic==2.13.4
urllib3==2.6.3
